[백준] 2749 피보나치 수 3
시간 제한 메모리 제한 제출 정답 맞은 사람 정답 비율
1 초 128 MB 18017 5857 4740 38.424%
문제
피보나치 수는 0과 1로 시작한다. 0번째 피보나치 수는 0이고, 1번째 피보나치 수는 1이다. 그 다음 2번째 부터는 바로 앞 두 피보나치 수의 합이 된다.

이를 식으로 써보면 Fn = Fn-1 + Fn-2 (n ≥ 2)가 된다.

n=17일때 까지 피보나치 수를 써보면 다음과 같다.

0, 1, 1, 2, 3, 5, 8, 13, 21, 34, 55, 89, 144, 233, 377, 610, 987, 1597

n이 주어졌을 때, n번째 피보나치 수를 구하는 프로그램을 작성하시오.

입력
첫째 줄에 n이 주어진다. n은 1,000,000,000,000,000,000보다 작거나 같은 자연수이다.

출력
첫째 줄에 n번째 피보나치 수를 1,000,000으로 나눈 나머지를 출력한다.

예제 입력 1
1000
예제 출력 1
228875

+++ src/B_2749/B_2749.kt
@@ -8,7 +8,7 @@ fun main(args: Array<String>)=with(BufferedReader(InputStreamReader(System.`in`)
     fi[1]=1
     var n:Long = readLine().toLong()
     for(i in 2..1_500_000){
-        fi[i.toInt()] = (fi[(i-1).toInt()]+fi[(i-2).toInt()])%1_000_000
+        fi[i] = (fi[(i-1)]+fi[(i-2)])%1_000_000
     }
     if(n >= 1_500_000)
         n%=1_500_000
